外來鍵欄位屬性修改

+++ database/migrations/2021_03_26_084820_create_comments_table.php
@@ -15,8 +15,8 @@ class CreateCommentsTable extends Migration
     {
         Schema::create('comments', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('trail_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('trail_id');
             $table->date('date');
             $table->integer('star');
             $table->integer('difficulty');
